More refactor/cleanup
